fill in proof block at the end of scope construction
